requirements.txt
# Install with: uv pip install -r requirements.txt
pandas>=2.0.0
yfinance>=0.2.40
fredapi>=0.5.1
python-dotenv>=1.0.0
numpy>=1.26.0
matplotlib>=3.8.0
seaborn>=0.13.0
jupyter>=1.0.0
ipykernel>=6.29.0
